ORD-PROD-4: Refactor order-service

+++ services/product-service/internal/domain/product.go
@@ -15,7 +15,7 @@ type Product struct {
 	Description  string `bun:"description"`
 	Price        int64 `bun:"price,notnull"`
 	Category     string `bun:"category,notnull"`
-	Rating       float64 `bun:"rating,notnull"`
+	Rating       float32 `bun:"rating,notnull"`
 	DeliveryDays int `bun:"delivery_days,notnull"`
 	CreatedAt    time.Time `bun:"created_at,notnull"`
 	UpdatedAt    time.Time `bun:"updated_at,notnull"`
